Separate app and server
