+++ backend/server.js
@@ -1,35 +1,13 @@
-const express = require("express");
-const cors = require("cors");
+const app = require("./app");
+const dotenv = require("dotenv");
+dotenv.config();
 
-const authRoutes = require("./routes/auth");
 const connectToMongo = require("./db");
-
-
 connectToMongo();
-
-const app = express();
-
-var corsOptions = {
-  origin: "http://localhost:8081"
-};
-
-app.use(cors(corsOptions));
-
-// parse requests of content-type - application/json
-app.use(express.json());
-
-// parse requests of content-type - application/x-www-form-urlencoded
-app.use(express.urlencoded({ extended: true }));
-
-// simple route
-app.get("/", (req, res) => {
-  res.json({ message: "Hello PU" });
-});
-
-app.use('/auth', authRoutes);
 
 // set port, listen for requests
 const PORT = process.env.PORT || 8080;
+
 app.listen(PORT, () => {
   console.log(`Server is running on port ${PORT}.`);
 });
